requirements.txt
# Core Dependencies
Kivy==2.3.1
Kivy-Garden==0.1.5
python-dotenv==1.1.1
Pygments==2.19.2

# Market Data & APIs
yfinance==0.2.66
alpaca-py==0.42.2
finnhub-python==2.4.25

# Data Processing
pandas==2.3.2
numpy==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Web & HTTP
requests==2.32.5
urllib3==2.5.0
beautifulsoup4==4.14.0
soupsieve==2.8
curl_cffi==0.13.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10

# WebSockets
websocket-client==1.8.0
websockets==15.0.1

# SSE (Server-Sent Events)
sseclient-py==1.8.0

# Serialization & Data
msgpack==1.1.1
protobuf==6.32.1

# Database
peewee==3.18.2

# Type Checking & Validation
pydantic==2.11.9
pydantic_core==2.33.2
annotated-types==0.7.0
typing-inspection==0.4.2
typing_extensions==4.15.0

# Utilities
docutils==0.22.2
filetype==1.2.0
frozendict==2.4.6
multitasking==0.0.12
platformdirs==4.4.0
six==1.17.0

# Low-level
cffi==2.0.0
pycparser==2.23